Move StatesQuery into Country namespace
This align the query namespace with other query objects
