+++ spec/lib/spree/queries/country/states_query_spec.rb
@@ -2,7 +2,7 @@
 
 require 'spec_helper'
 
-RSpec.describe Spree::Queries::StatesQuery do
+RSpec.describe Spree::Queries::Country::StatesQuery do
   let(:country) { create(:country) }
 
   let!(:states) { create_list(:state, 2, country: country) }
